refactor: follow ai rule
