+++ src/pages/logExplorer/components/LogsViewer/Clustering/Table.tsx
@@ -1,22 +1,19 @@
 // 日志聚类，目前仅支持了doris和es
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Space, Tag, Dropdown, Button, Menu, Popover, Spin, Progress, Row, Col, Statistic, Tooltip } from 'antd';
 import { CaretDownOutlined, MinusCircleOutlined, PlusCircleOutlined, CopyOutlined, BarChartOutlined, SyncOutlined } from '@ant-design/icons';
-import { copy2ClipBoard } from '@/utils';
 import { createPortal } from 'react-dom';
 import { useTranslation } from 'react-i18next';
 import _ from 'lodash';
-import { PRIMARY_COLOR } from '@/utils/constant';
 import { NAME_SPACE } from '../../../constants';
 import { getLogClustering, ClusteringItem, getQueryClustering, getLogPattern } from '../../../services';
-import { ClusterPattern } from '../../../types';
 import { getGlobalConfig } from '@/plus/components/LogDownload/service';
 import { OnValueFilterParams, OptionsType } from '../types';
 import { Field } from '@/plugins/doris/ExplorerNG/types';
 import RDGTable from '../components/Table';
-const DEFAULT_MAX_LOG_COUNT = 10000000;
+const DEFAULT_MAX_LOG_COUNT = 100000;
 import { DatasourceCateEnum } from '@/utils/constant';
-import ExistsIcon from '@/pages/explorer/components/RenderValue/ExistsIcon';
+import { PatternPopover, ConstPopover } from './Popover';
 interface Props {
   onValueFilter: (condition: OnValueFilterParams) => void;
   indexData: Field[];
@@ -82,164 +79,6 @@ export default function TableCpt(props: Props) {
       });
   };
 
-  const PatternPopover = ({ uuid, partId, children, title }: { uuid: string; partId: number; children: React.ReactNode; title: string }) => {
-    const [visible, setVisible] = useState(false);
-    const [loading, setLoading] = useState(false);
-    const [patternData, setPatternData] = useState<ClusterPattern | null>(null);
-
-    const handleVisibleChange = (v: boolean) => {
-      setVisible(v);
-      if (v && !patternData) {
-        setLoading(true);
-        getLogPattern(uuid, partId)
-          .then((res) => {
-            setPatternData(res);
-          })
-          .finally(() => {
-            setLoading(false);
-          });
-      }
-    };
-
-    const content = (
-      <div style={{ width: 320 }}>
-        <Spin spinning={loading}>
-          {patternData && (
-            <>
-              <div className='bg-fc-200 p-4'>
-                <Row gutter={[16, 16]}>
-                  <Col span={12}>
-                    <Statistic className='n9e-logexplorer-field-statistic text-center' title={t('stats.unique_count')} value={patternData.count} />
-                  </Col>
-                  <Col span={12}>
-                    <Statistic className='n9e-logexplorer-field-statistic text-center' title={t('stats.exist_ratio')} value={patternData.percentage} suffix='%' />
-                  </Col>
-                </Row>
-              </div>
-              <div>
-                <div className='my-2 text-l2'>
-                  <strong>{t('clustering.top5_title')}</strong>
-                </div>
-                {_.isEmpty(patternData.top5) && t('clustering.no_data')}
-                {_.map(_.orderBy(patternData.top5, ['count'], ['desc']), (item) => {
-                  const percent = _.floor(item.percentage, 2);
-                  return (
-                    <div key={item.value} className='mb-2'>
-                      <div className='flex justify-between'>
-                        <Tooltip title={item.value}>
-                          <div style={{ width: 'calc(100% - 80px)' }} className='truncate'>
-                            {_.isEmpty(item.value) && !_.isNumber(item.value) ? '(empty)' : item.value}
-                          </div>
-                        </Tooltip>
-                        <span>{item.count?.toLocaleString()}</span>
-                      </div>
-                      <div className='flex justify-between items-center'>
-                        <div style={{ width: 'calc(100% - 60px)' }}>
-                          <Progress percent={percent} size='small' showInfo={false} strokeColor={PRIMARY_COLOR} />
-                        </div>
-                        <span>{percent}%</span>
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-            </>
-          )}
-        </Spin>
-      </div>
-    );
-
-    return (
-      <Popover trigger='click' visible={visible} onVisibleChange={handleVisibleChange} content={content} title={title}>
-        {children}
-      </Popover>
-    );
-  };
-
-  const ConstPopover = ({ value, children }: { value: string; children: React.ReactNode }) => {
-    const [popoverVisible, setPopoverVisible] = useState(false);
-    return (
-      <Popover
-        visible={popoverVisible}
-        onVisibleChange={setPopoverVisible}
-        trigger='click'
-        overlayClassName='explorer-origin-field-val-popover'
-        content={
-          <ul className='ant-dropdown-menu ant-dropdown-menu-root ant-dropdown-menu-vertical ant-dropdown-menu-light'>
-            <li
-              className='ant-dropdown-menu-item ant-dropdown-menu-item-only-child'
-              onClick={() => {
-                setPopoverVisible(false);
-                copy2ClipBoard(`${field}:${value}`);
-              }}
-            >
-              <Space>
-                <CopyOutlined />
-                {t('common:btn.copy')}
-              </Space>
-            </li>
-            <li
-              className='ant-dropdown-menu-item ant-dropdown-menu-item-only-child'
-              onClick={() => {
-                setPopoverVisible(false);
-                props.onValueFilter({
-                  key: field,
-                  value,
-                  assignmentOperator: '=',
-                  operator: 'AND',
-                });
-              }}
-            >
-              <Space>
-                <PlusCircleOutlined />
-                {t('logs.filterAllAnd')}
-              </Space>
-            </li>
-            <li
-              className='ant-dropdown-menu-item ant-dropdown-menu-item-only-child'
-              onClick={() => {
-                setPopoverVisible(false);
-                props.onValueFilter({
-                  key: field,
-                  value,
-                  assignmentOperator: '=',
-                  operator: 'NOT',
-                });
-              }}
-            >
-              <Space>
-                <MinusCircleOutlined />
-                {t('logs.filterAllNot')}
-              </Space>
-            </li>
-            {cate === DatasourceCateEnum.elasticsearch && (
-              <li
-                className='ant-dropdown-menu-item ant-dropdown-menu-item-only-child'
-                onClick={() => {
-                  setPopoverVisible(false);
-                  props.onValueFilter?.({
-                    key: field,
-                    value: value,
-                    assignmentOperator: '=',
-                    operator: 'EXISTS',
-                    indexName: field,
-                  });
-                }}
-              >
-                <Space>
-                  <ExistsIcon />
-                  {t('logs.filterExists')}
-                </Space>
-              </li>
-            )}
-          </ul>
-        }
-      >
-        {children}
-      </Popover>
-    );
-  };
-
   const getColumns = () => {
     const columns: any[] = [
       {
@@ -250,12 +89,12 @@ export default function TableCpt(props: Props) {
         key: 'parts',
         name: t('clustering.log_data'),
         formatter: ({ row }) => {
-          const sortedParts = row.parts?.sort((a, b) => {
+          const sortedParts = [...(row.parts ?? [])].sort((a, b) => {
             if (a.type === 'pattern' && b.type !== 'pattern') return -1;
             if (a.type !== 'pattern' && b.type === 'pattern') return 1;
             return 0;
           });
-          return <Space size={'small'}>{sortedParts?.map((part) => {
+          return <Space size={'small'}>{sortedParts.map((part) => {
             if (part.type === 'pattern') {
               if (backEndCluster) {
                 return (
@@ -269,7 +108,7 @@ export default function TableCpt(props: Props) {
             } else {
               if (backEndCluster) {
                 return (
-                  <ConstPopover key={part.part_id} value={part.data}>
+                  <ConstPopover key={part.part_id} value={part.data} field={field} cate={cate} onValueFilter={props.onValueFilter}>
                     <span className='cursor-pointer hover:underline'>{part.data}</span>
                   </ConstPopover>
                 );
